\name{heterometa-package}
\alias{heterometa-package}
\alias{heterometa}
\docType{package}
\title{\packageTitle{heterometa}}
\description{\packageDescription{heterometa}}
\details{
\packageIndices{heterometa}

Published meta--analyses often only provide
a subset of the
heterogeneity measures first described
in \insertCite{higgins02;textual}{heterometa}.
This package provides a way of converting between
some of them.
It allows for the input of I2, Q or p
and outputs Q, H, I2, and confidence intervals
for them.
It also contains the data used as examples
in that original article.

}
%\author{\packageAuthor{heterometa}}

%\packageMaintainer{heterometa}

\references{
\insertAllCited{}
}
\keyword{ package }
